Physical blocks are off_t's, not uint32's.
git-svn-id: file:///srv/svn/repos/haiku/trunk/current@5314 a95241bf-73f2-0310-859d-f6bbb57e9c96

+++ src/add-ons/kernel/file_systems/udf/MetadataPartition.cpp
@@ -30,7 +30,7 @@ MetadataPartition::~MetadataPartition()
 /*! \brief Maps the given logical block to a physical block on disc.
 */
 status_t
-MetadataPartition::MapBlock(uint32 logicalBlock, uint32 &physicalBlock)
+MetadataPartition::MapBlock(uint32 logicalBlock, off_t &physicalBlock)
 {
 	return B_NOT_IMPLEMENTED;
 }
